requirements.txt
streamlit
pdfplumber
ollama
google-genai